feat: restructure header and footer components into global directory
